### 2021-12-17 #4007 @whedon
[PRE REVIEW]: Magphi: Sequence extraction tool from FASTA and GFF3 files using seed pairs
**Submitting author:** @milnus (<a href="http://orcid.org/0000-0001-9751-9877">Magnus Jespersen</a>)
**Repository:** https://github.com/milnus/Magphi
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.6
