### 2022-04-24 #4007 @majensen
Hi @huddlej  -- would you have time and inclination to review [this work](#4007) for [JOSS](https://joss.theoj.org)? It would be great to have your microbe sequencing expertise here. Thanks for your consideration --@majensen
