### 2022-04-26 #4007 @majensen
Hey @amoeba - have you got cycles to review [this work](#4007) for [JOSS](https://joss.theoj.org)? Consideration much appreciated - @majensen
