### 2022-04-27 #4007 @majensen
Hi @gtonkinhill - Would you be willing to review [this software submission](#4007) for our [Journal of Open Source Software (JOSS)](https://joss.theoj.org)? Please have a look at our [guide for reviewers](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html), and let me know here in this thread, or via email to me at maj \-dot\- fortinbras \-at\- gmail \-dot\- com. Your expertise here would be invaluable.
Really appreciate your consideration    --@majensen
