### 2022-04-27 #4007 @majensen
@gtonkinhill thanks very much! I appreciate the transparency and I don't see any issues.
(P.S., in olden times I used interact on occasion with your old boss Simon when I was working in HIV. Apropos of not much except what a small world this is...)
