### 2022-05-02 #4007 @majensen
Hey @dcnickle my friend - would you have time to review [this work](#4007) for [JOSS](https://joss.theoj.org)? I really appreciate the consideration.
