### 2022-05-02 #4007 @dcnickle
Yes - I can review.

David

> On May 2, 2022, at 10:56 AM, Mark Jensen ***@***.***> wrote:
>
>
> Hey @dcnickle <https://github.com/dcnickle> my friend - would you have time to review this work <x-msg://12/#4007> for JOSS <https://joss.theoj.org/>? I really appreciate the consideration.
>
