### 2022-05-06 milnus/Magphi#11 @gtonkinhill
[JOSS Review] - Conda installation is not yet completed
[Review](https://github.com/openjournals/joss-reviews/issues/4369)

It would be good to make Magphi available on bioconda. This is mentioned in the documentation but not yet completed.
