### 2022-05-06 milnus/Magphi#12 @gtonkinhill
[JOSS Review] Dependencies need to be listed
[Review](https://github.com/openjournals/joss-reviews/issues/4369)

Currently, the dependencies are only listed as part of a conda command. I think JOSS requires that these are given as a separate section in the documentation.
