### 2022-05-06 milnus/Magphi#11 @gtonkinhill
The recipe for Bakta might be helpful https://github.com/bioconda/bioconda-recipes/blob/master/recipes/bakta/meta.yaml
