### 2022-05-06 milnus/Magphi#14 @gtonkinhill
[JOSS Review] Tests are currently failing
[Review = https://github.com/openjournals/joss-reviews/issues/4369]

I am not sure if the tests have actually failed or the tag in the GitHub readme has just failed to update. In either case, it would be good to re-run the tests so that the tag updates to green.
