### 2022-05-06 milnus/Magphi#15 @gtonkinhill
[JOSS Review] Community guidelines required
[Review](https://github.com/openjournals/joss-reviews/issues/4369)

JOSS requires that there is a community guidelines section in the documentation.

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
