### 2022-05-06 milnus/Magphi#16 @gtonkinhill
[JOSS Review] Spelling and grammar
[Review](https://github.com/openjournals/joss-reviews/issues/4369)

The paper is well written, however, there are quite a few typos in the Readme and Wiki. There are GitHub actions to check spelling, but I think it might be easiest to simply copy the `.md` files into a word editor such as Google docs to help clean up the remaining typos.

A few I've spotted include:

- Readme and wiki: 'Examlpe', 'seqeunces', 'imporive' ...
- Wiki - installation: 'seqeunce', 'Qucik', 'writen', 'Comming', 'recipie', ...
