### 2022-05-06 #4007 @dcnickle
I need this to be assigned to @DavidNickle.

@dcnickle is broken

David

> On May 3, 2022, at 4:17 PM, The Open Journals editorial robot ***@***.*** ***@***.***>> wrote:
>
>
> Assigned #4007 <https://github.com/openjournals/joss-reviews/issues/4007> to @dcnickle <https://github.com/dcnickle>.
>
