### 2022-05-08 f679f16
corrected spelling in readme doc
spelling and grammer corrections to address JOSS review comments
