+++ README.md
@@ -12,7 +12,7 @@ may not be ideal due to the inability for BLAST to deal with short sequences. co
 
 ## Installation
 ### Conda installation
-``` # Comming soon ```
+``` # Coming soon ```
 
 ### Pip install
 Make sure you have the right versions of Samtools and Bedtools installed
@@ -20,11 +20,11 @@ Make sure you have the right versions of Samtools and Bedtools installed
 ``` pip install Magphi```
 
 ### Setting up conda environment using Conda 
-Solution until conda recipie has been created  
+Solution until conda recipe has been created  
 ```$ conda create -n Magphi -c bioconda samtools=1.13 blast bedtools python=3.9```  
 ```$ conda activate Magphi```  
 ```$ pip install Magphi```  
-To test the insatllation of dependencies run:  
+To test the installation of dependencies run:  
 ```$ Magphi --check```
 
 ## Help command
@@ -38,7 +38,7 @@ optional arguments:
   -g .fa/.gff [.fa/.gff ...], --input_genomes .fa/.gff [.fa/.gff ...]
                         Give the fasta or gff3 files. (gff3 files should contain the genome fasta sequence)
   -s multi_fasta_file.fa, --input_seeds multi_fasta_file.fa
-                        Give the multi fasta containing the seed sequences to be used for extracting sequnces
+                        Give the multi fasta containing the seed sequences to be used for extracting sequences
   -is, --include_seeds  Argument to include the seeds in the sequence/annotations extracted [default: seeds are removed]
   -md int, --max_seed_distance int
                         The maximum distance with which seeds will be merged. This can often be set a bit higher than an expected size of a region If no maximum distance is wanted then set to 0 [default: 50,000bp]
@@ -54,15 +54,15 @@ optional arguments:
   ```
 
 ## Constructing seed sequence file
-Seed sequences are what define a region in which the user is interested. These seqeunces can be core genes, or known stretches of DNA that surround a specific region in genomes. It is recommended that seed seqeunces are no less than 100 bp, as BLAST does not work well for smaller pieces of DNA.
+Seed sequences are what define a region in which the user is interested. These sequences can be core genes, or known stretches of DNA that surround a specific region in genomes. It is recommended that seed sequences are no less than 100 bp, as BLAST does not work well for smaller pieces of DNA.
 
-### Rules to follow for seed sequnces
-* Seed seqeunces should be no less than 100 bp
-* Names should be unique for each pair of seed sequnces
+### Rules to follow for seed sequences
+* Seed sequences should be no less than 100 bp
+* Names should be unique for each pair of seed sequences
 * Each name of a seed sequence in a pair should be appended with a unique tag (i.e. ```_1``` or ```_2```) to make them unique within the pair.
-* If a seed seqeunces is used twice it should be duplicated in the fasta file and given two unique names
+* If a seed sequences is used twice it should be duplicated in the fasta file and given two unique names
 
-### Examlpe seed sequence file:
+### Example seed sequence file:
 ```
 >phageA_1
 CTGGTCGGTTGGTATATAGTGTTCGCACTG.....
@@ -75,7 +75,7 @@ TGCAGTTCCCCATGGGGTGTAGGTGTAACG.....
 ```
 
 ## Genomes
-Genomes should be in either Fasta format or GFF3 format with the genome appended in the end of the file, with the ##FASTA line seperating the annotations from the appended fasta genome.
+Genomes should be in either Fasta format or GFF3 format with the genome appended in the end of the file, with the ##FASTA line separating the annotations from the appended fasta genome.
[…]
