+++ README.md
@@ -1,8 +1,152 @@
-[![Build Status](https://app.travis-ci.com/milnus/Magphi.svg?branch=main)](https://app.travis-ci.com/milnus/Magphi)
+[![Test](https://github.com/milnus/Magphi/actions/workflows/test.yml/badge.svg)](https://github.com/milnus/Magphi/actions/workflows/test.yml)
 
 # Magphi
-Tool intended to pull out sequences and annotations between insertion sites specified by the user.
-**THIS TOOL IS STILL IN DEVELOPMENT AND MAY CONTAIN BUGS🐛! IF YOU FIND A BUG PLEASE RAISE AN ISSUE AND WE WILL FIX IT AS QUICKLY AS WE CAN**
+**Tool intended to pull out sequences and annotations from genomes between user provided fasta sequence paris (seed sequences) using BLAST**
 
-See Wiki tab for more info on the workings, dependencies and outputs of the tool
+## When to use
+if you have moderately large (100+bp) sequences such as genes that you know flank a particular region of your genome
+such as prophage attachment sites and wish to extract the region in one or more genomes. Particularly useful if there is
+some level of acceptable variation in the site that will allow a hit with blast but not a direct pattern match.
+If you have small (<100 bp) motifs with confidence as to the number of ambiguities and direction of primer this tool
+may not be ideal due to the inability for BLAST to deal with short sequences. consider trying [Seqkit](https://bioinf.shenwei.me/seqkit/) instead.
 
+## Installation
+#### Dependencies
+* Python >= 3.9
+* Blast
+* Bedtools
+* Samtools
+* pybedtools
+### Conda installation
+``` # Coming soon ```
+
+### Pip install
+Make sure you have the right versions of Samtools and Bedtools installed
+
+``` pip install Magphi```
+
+### Setting up conda environment using Conda 
+Solution until conda recipe has been created  
+```$ conda create -n Magphi -c bioconda samtools=1.13 blast bedtools python=3.9```  
+```$ conda activate Magphi```  
+```$ pip install Magphi```  
+To test the installation of dependencies run:  
+```$ Magphi --check```
+
+## Help command
+```
+usage: Magphi [-h] -g .fa/.gff [.fa/.gff ...] -s multi_fasta_file.fa [-is] [-md int] [-b | -n] [-o path/to/output] [-c int] [-l | -q] [--check] [-v]
+
+Welcome to Magphi! This program will extract sequences and possible annotations within a set of seed sequences given as input.
+
+optional arguments:
+  -h, --help            show this help message and exit
+  -g .fa/.gff [.fa/.gff ...], --input_genomes .fa/.gff [.fa/.gff ...]
+                        Give the fasta or gff3 files. (gff3 files should contain the genome fasta sequence)
+  -s multi_fasta_file.fa, --input_seeds multi_fasta_file.fa
+                        Give the multi fasta containing the seed sequences to be used for extracting sequences
+  -is, --include_seeds  Argument to include the seeds in the sequence/annotations extracted [default: seeds are removed]
+  -md int, --max_seed_distance int
+                        The maximum distance with which seeds will be merged. This can often be set a bit higher than an expected size of a region If no maximum distance is wanted then set to 0 [default: 50,000bp]
+  -b, --print_breaks    Argument to print outputs when seeds are next to contig breaks [default: sequences are not printed]
+  -n, --no_sequences    Argument to not print outputs related to annotations or sequences found between seeds [default: sequences are printed]
+  -o path/to/output, --output_folder path/to/output
+                        Give path to output folder [default: current folder]
+  -c int, --cpu int     Give max number of CPUs [default: 1]
+  -l, --log             Record program progress in for debugging purpose
+  -q, --quiet           Only print warnings
[…]
