### 2022-05-10 3a8510d
Add in the Conda installation instructions and badge
