### 2022-05-26 milnus/Magphi#16 @milnus
Both the Readme and wiki have been addressed in #17
