### 2022-05-26 milnus/Magphi#15 @milnus
Addressed in [dev's ReadMe](https://github.com/milnus/Magphi/tree/dev#community-contributions-via-github)
