### 2022-05-26 milnus/Magphi#14 @milnus
This was a product of a manually cancel test. The badge is up to date and shows test passing
