### 2022-06-23 milnus/Magphi#11 @milnus
Magphi is now on Bioconda and can be installed using: ```conda install -c bioconda magphi```
