### 2022-06-28 milnus/Magphi#12 @gtonkinhill
This has been addressed in the updated documentation.
