### 2022-06-28 milnus/Magphi#11 @gtonkinhill
Great! I was able to install Magphi using conda. Could you update the readme with this option?
