### 2022-06-29 ee2c893
Add Conda install to README
