+++ README.md
@@ -16,6 +16,10 @@ may not be ideal due to the inability for BLAST to deal with short sequences. co
 ### Conda installation
 ``` conda install -c bioconda magphi ```
 
+### Pip install
+Make sure you have the right versions of Samtools and Bedtools installed
+``` pip install Magphi```
+
 #### Dependencies
 * Python >= 3.9
 * Blast
@@ -23,9 +27,6 @@ may not be ideal due to the inability for BLAST to deal with short sequences. co
 * Samtools
 * pybedtools
 
-### Pip install
-Make sure you have the right versions of Samtools and Bedtools installed
-``` pip install Magphi```
 
 ## Help command
 ```
