### 2022-06-30 #4369 @majensen
@milnus @gtonkinhill -- my apologies for neglecting this submission. Are there any items outstanding regarding gtonkinhill's review?

@DavidNickle can you provide your review via the [checklist above](https://github.com/openjournals/joss-reviews/issues/4369#issuecomment-1171613322)?
