### 2022-07-04 #4369 @majensen
Hi @milnus - there are only two minor items: 1) missing a period at the end of sentence, l. 37 in proof, 2) I think the Mount ref. is fine, but could you also add the original BLAST algorithm paper (https://pubmed.ncbi.nlm.nih.gov/2231712/). Just my preference.

Once you make and push any changes, can you please create an open archive of the code repo at [Zenodo](https://zenodo.org), [FigShare](https://figshare.com) or similar, and report the DOI back to this thread? [Here is a simple tutorial](https://github.com/openjournals/joss-reviews/issues/1839#issuecomment-560048729) for using Zenodo.
