### 2022-07-10 5a5bb04
Add in changes to paper given in review
