+++ paper/paper.bib
@@ -22,6 +22,17 @@
   publisher={Cold Spring Harbor Laboratory Press}
 }
 
+@article{altschul:1990,
+  title={Basic local alignment search tool},
+  author={Altschul, Stephen F and Gish, Warren and Miller, Webb and Myers, Eugene W and Lipman, David J},
+  journal={Journal of molecular biology},
+  volume={215},
+  number={3},
+  pages={403--410},
+  year={1990},
+  publisher={Elsevier}
+}
+
 @article{shen:2016,
     doi = {10.1371/journal.pone.0163962},
     author = {Shen, Wei AND Le, Shuai AND Li, Yan AND Hu, Fuquan},
