### 2022-07-10 c29f9a1
Add secound reference for original BLAST algorithm
