+++ paper/paper.bib
@@ -30,6 +30,7 @@
   number={3},
   pages={403--410},
   year={1990},
+  doi={10.1016/S0022-2836(05)80360-2}
   publisher={Elsevier}
 }
 
