### 2022-07-12 #4369 @arfon
@milnus – looks like that last chance to the BibTeX was missing a comma. I've fixed it in here: https://github.com/milnus/Magphi/pull/19
